+++ src/_pytask/click.py
@@ -108,6 +108,9 @@ def print_options(group_or_command: click.Command | DefaultGroup, ctx: Any) -> N
         if isinstance(param, click.Argument):
             continue
 
+        if getattr(param, "hidden", False):
+            continue
+
         # The ordering of -h and --help is not fixed.
         if param.name == "help":
             opt1 = highlighter("-h")
@@ -115,12 +118,18 @@ def print_options(group_or_command: click.Command | DefaultGroup, ctx: Any) -> N
         elif len(param.opts) == 2:
             opt1 = highlighter(param.opts[0])
             opt2 = highlighter(param.opts[1])
+        elif len(param.opts) == 1 and len(param.secondary_opts) == 1:
+            opt1 = Text("")
+            opt2 = highlighter(param.opts[0] + "/" + param.secondary_opts[0])
         else:
             opt1 = Text("")
             opt2 = highlighter(param.opts[0])
 
         if param.metavar:
             opt2 += Text(f" {param.metavar}", style="metavar")
+        elif isinstance(param.type, click.Choice):
+            choices = "[" + "|".join(param.type.choices) + "]"
+            opt2 += Text(f" {choices}", style="metavar", overflow="fold")
 
         help_record = param.get_help_record(ctx)
         if help_record is None: