+++ src/_pytask/database.py
@@ -5,14 +5,23 @@ from enum import Enum
 from pathlib import Path
 from typing import Any
 
-import attr
 import click
 import networkx as nx
 from _pytask.config import hookimpl
+from _pytask.config_utils import parse_click_choice
 from _pytask.dag import node_and_neighbors
 from _pytask.shared import convert_truthy_or_falsy_to_bool
 from _pytask.shared import get_first_non_none_value
+from _pytask.typed_settings import option
 from pony import orm
+
+
+class _DatabaseProviders(Enum):
+    SQLITE = "sqlite"
+    POSTGRES = "postgres"
+    MYSQL = "mysql"
+    ORACLE = "oracle"
+    COCKROACH = "cockroach"
 
 
 db = orm.Database()
@@ -68,42 +77,29 @@ class DatabaseProviders(Enum):
 @hookimpl
 def pytask_extend_command_line_interface(cli: click.Group) -> None:
     """Extend command line interface."""
-    cli["build"]["options"]["database_provider"] = attr.ib(
-        default=DatabaseProviders.sqlite, type=DatabaseProviders
+    cli["build"]["options"]["database_provider"] = option(
+        default=DatabaseProviders.sqlite,
+        type=DatabaseProviders,
+        help=(
+            "Database provider. All providers except sqlite are considered "
+            "experimental. [dim]\\[default: sqlite][/]"
+        ),
     )
-    cli["build"]["options"]["database_filename"] = attr.ib(default=None, type=str)
-    cli["build"]["options"]["database_create_db"] = attr.ib(default=True, type=bool)
-    cli["build"]["options"]["database_create_tables"] = attr.ib(default=True, type=bool)
-    # additional_parameters = [
-    #     click.Option(
-    #         ["--database-provider"],
-    #         type=click.Choice(["sqlite", "postgres", "mysql", "oracle", "cockroach"]),
-    #         help=(
-    #             "Database provider. All providers except sqlite are considered "
-    #             "experimental.  [default: sqlite]"
-    #         ),
-    #         default=None,
-    #     ),
-    #     click.Option(
-    #         ["--database-filename"],
-    #         type=click.Path(),
-    #         help="Path to database relative to root.  [default: .pytask.sqlite3]",
-    #         default=None,
-    #     ),
-    #     click.Option(
-    #         ["--database-create-db"],
-    #         type=bool,
-    #         help="Create database if it does not exist.  [default: True]",
-    #         default=None,
-    #     ),
-    #     click.Option(
-    #         ["--database-create-tables"],
-    #         type=bool,
-    #         help="Create tables if they do not exist.  [default: True]",
-    #         default=None,
-    #     ),
-    # ]
-    # cli.commands["build"].params.extend(additional_parameters)
+    cli["build"]["options"]["database_filename"] = option(
+        default=None,
+        type=str,
+        help="Path to database relative to root. [dim]\\[default: .pytask.sqlite3][/]",
+    )
+    cli["build"]["options"]["database_create_db"] = option(
+        default=True,
+        type=bool,
+        help="Create database if it does not exist. [dim]\\[default: True][/]",
+    )
+    cli["build"]["options"]["database_create_tables"] = option(
+        default=True,
+        type=bool,
+        help="Create tables if they do not exist. [dim]\\[default: True][/]",
+    )
 
 
 @hookimpl
@@ -114,7 +110,11 @@ def pytask_parse_config(
 ) -> None:
     """Parse the configuration."""
     config["database_provider"] = get_first_non_none_value(
-        config_from_cli, config_from_file, key="database_provider", default="sqlite"
+        config_from_cli,
+        config_from_file,
+        key="database_provider",
+        default=_DatabaseProviders.SQLITE,
+        callback=parse_click_choice("database_provider", _DatabaseProviders),
     )
     filename = get_first_non_none_value(
         config_from_cli,
@@ -142,7 +142,7 @@ def pytask_parse_config(
         callback=convert_truthy_or_falsy_to_bool,
     )
     config["database"] = {
-        "provider": config["database_provider"],
+        "provider": config["database_provider"].value,
         "filename": config["database_filename"].as_posix(),
         "create_db": config["database_create_db"],
         "create_tables": config["database_create_tables"],