+++ src/_pytask/live.py
@@ -17,6 +17,7 @@ from _pytask.outcomes import TaskOutcome
 from _pytask.report import CollectionReport
 from _pytask.report import ExecutionReport
 from _pytask.shared import get_first_non_none_value
+from _pytask.typed_settings import option
 from rich.live import Live
 from rich.status import Status
 from rich.table import Table
@@ -26,16 +27,12 @@ from rich.text import Text
 @hookimpl
 def pytask_extend_command_line_interface(cli: click.Group) -> None:
     """Extend command line interface."""
-    cli["build"]["options"]["n_entries_in_table"] = attr.ib(type=int, default=15)
-    # additional_parameters = [
-    #     click.Option(
-    #         ["--n-entries-in-table"],
-    #         default=None,
-    #         help="How many entries to display in the table during the execution. "
-    #         "Tasks which are running are always displayed.  [default: 15]",
-    #     ),
-    # ]
-    # cli.commands["build"].params.extend(additional_parameters)
+    cli["build"]["options"]["n_entries_in_table"] = option(
+        type=int,
+        default=15,
+        help="How many entries to display in the table during the execution. "
+        "Tasks which are running are always displayed. [dim]\\[default: 15][/]",
+    )
 
 
 @hookimpl